requirements.txt
stable-baselines3[extra]>=2.6.0
gymnasium[mujoco]>=1.1.0
mujoco>=3.2.0
PyYAML>=6.0.2
tensorboard>=2.18.0
